+++ sdk/eventgrid/azure-messaging-eventgrid/src/main/java/com/azure/messaging/eventgrid/systemevents/AvsClusterEventData.java
@@ -14,8 +14,12 @@ import java.util.List;
 
 /**
  * Schema of the Data property of an EventGridEvent for Microsoft.AVS/clusters events.
+ * 
+ * @deprecated This class is deprecated and may be removed in future releases. System events are now available in the
+ * azure-messaging-eventgrid-systemevents package.
  */
 @Fluent
+@Deprecated
 public class AvsClusterEventData implements JsonSerializable<AvsClusterEventData> {
 
     /*