+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### Other Changes
 
+## 12.23.1 (2025-06-19)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-storage-blob` from `12.30.0` to version `12.30.1`.
+- Upgraded `azure-core-http-netty` from `1.15.11` to version `1.15.12`.
+- Upgraded `azure-core` from `1.55.3` to version `1.55.4`.
+
 ## 12.24.0-beta.1 (2025-05-06)
 
 ### Features Added